pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "ai-workflow-engine"
version = "2.0.0"
description = "AI-Native Workflow Engine - A structured workflow management tool for AI-assisted development"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [
    {name = "Workflow Tool Team"}
]
keywords = ["workflow", "ai", "development", "automation"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
dependencies = [
    "PyYAML>=6.0",
]

[project.scripts]
flow = "workflow.cli:main"

[project.urls]
Homepage = "https://github.com/workflow-tool/workflow-tool"
Documentation = "https://github.com/workflow-tool/workflow-tool#readme"
Repository = "https://github.com/workflow-tool/workflow-tool"

[tool.setuptools.packages.find]
where = ["."]
include = ["workflow*"]
